refactor(metadata): Extract user operations to user.rs
Extract 28 user-related database methods from db.rs to a new user.rs module:

- User initialization (4 methods)
- User CRUD operations (7 methods)
- Subscription management (4 methods)
- Session management (4 methods)
- Token management (5 methods)
- User preferences (8 methods)

Part of #51
Closes #54
