+++ backend/src/metadata/mod.rs
@@ -1,6 +1,7 @@
 mod db;
 mod pool;
 mod types;
+mod user;
 
 pub use pool::MetadataDb;
 pub use types::*;
